MultiChannel Audio Support
API remains same for existing AudioDevices, But if you need to use a custom audiodevice (with different channel support with OpenAL, this code supports 1,2,4,6,7 and 8 channel audio at 8 bit or 16 bit per sample resolution.

The default mode is that it works on the default audio device (tested on windows) unless you specify a particular device.